Adding article now works using all attributes available from DB. Shop (browse) page shows available articles

+++ src/main/java/irinakjoseva/vecnamoda/repository/ColorRepository.java
@@ -5,8 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ColorRepository extends JpaRepository<Color, Long> {
-//
-//    public List<Color> findAll();
+public interface ColorRepository extends JpaRepository<Color, Integer> {
 
 }
